add generic work/image routes
This switches our data files back to their original form,
for the most part (some names have been removed), using their
Hydra-generated IDs rather than `postcard` or `print`. This,
in tandem with `/build-catalog.json`, should allow adding items
to the 'collection' as simple as adding a JSON file to `public/data`
and running `npm run build-catalog`.

Also, as we're using this mostly for testing and, to my knowledge,
less-so for visual appeal, all files at `/downloads/*` resolve to
the same image (`/public/img/postcard.jpg`). If it seems like we'll
need to add separate images as well, we can look to update that
route to resolve to ID-named images in the same directory.