server rewrite/reorganizing (#7)
Verified as functional using a local development environment 👍 

* remove gulp, views, and unused dependencies

* rename app.js to index.js

* update static data and add catalog build script

* refactor generic routes

* display all debug logs

* allow catalog.json to have trailing slash

* use relative url for thumbnail_path

* fix how facets are built for catalog.json

* use generic ids for items

* fix usage typo

* move cli script to bin/

* move usage to bin/index.js

* add generic work/image routes

This switches our data files back to their original form,
for the most part (some names have been removed), using their
Hydra-generated IDs rather than `postcard` or `print`. This,
in tandem with `/build-catalog.json`, should allow adding items
to the 'collection' as simple as adding a JSON file to `public/data`
and running `npm run build-catalog`.

Also, as we're using this mostly for testing and, to my knowledge,
less-so for visual appeal, all files at `/downloads/*` resolve to
the same image (`/public/img/postcard.jpg`). If it seems like we'll
need to add separate images as well, we can look to update that
route to resolve to ID-named images in the same directory.

* add date facets to build, use new filenames

Just a few updates to the build script:

- adds a `date_artifact_{lower,upper}` facet to test date faceting
- removes previous `catalog.json` file (where present) at the beginning
  of the build
- imports all files from `public/data` for the catalog

* update readme

* fix typo in routes

* use absolute location for filepaths

* remove body-parser dependency